+++ api-bank/models/transaction.py
@@ -4,7 +4,7 @@ from services import db
 class Transaction(db.Model):
     __tablename__ = "transaction"
     id = db.Column(db.Integer, primary_key=True)
-    amount = db.Column(db.Numeric, nullable=False)
+    amount = db.Column(db.Float, nullable=False)
     date_created = db.Column(db.DateTime, nullable=False)
 
     account_id = db.Column(db.Integer, db.ForeignKey("account.id"))
